requirements.txt
python-telegram-bot
psutil
pymysql